Move draw setup code into VulkanBackend
